requirements.txt
appdirs==1.4.3
asgiref==3.2.7
CacheControl==0.12.6
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.0.7
django-allauth==0.42.0
django-crispy-forms==1.9.1
django-extensions==2.2.9
Faker==4.1.0
html5lib==1.0.1
idna==2.9
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
Pillow==7.1.2
progress==1.5
psycopg2==2.8.5
pyparsing==2.4.6
python-dateutil==2.8.1
python3-openid==3.1.0
pytoml==0.1.21
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
webencodings==0.5.1
